% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desaturate_cs.R
\name{desaturate_cs}
\alias{desaturate_cs}
\title{Destaturate a colour using any colour space}
\usage{
desaturate_cs(col = "#fedcba", cs = "oklch")
}
\arguments{
\item{col}{hex colour}

\item{cs}{coulur space}
}
\value{
hex colour
}
\description{
Destaturate a colour using any colour space
}
\examples{
desaturate_cs()
}
