\name{fastCircularPermutations-package}
\alias{fastCircularPermutations-package}
\docType{package}
\title{
 Very fast circular permutation analysis on binary outcomes
}
\description{
 File-Backed Matrix Class with Convenient Read and Write Access
}
\details{
	\tabular{ll}{
		Package: \tab filematrix \cr
		Type: \tab Package \cr
		License: \tab LGPL-3 \cr 
		Depends: \tab methods \cr
	}
	
	Provides very fast circular permutation analysis 
   using bitwise operations on binary representation of the data.
}
\note{
}
\author{
	Andrey A Shabalin \email{ashabalin@vcu.edu}
}
\keyword{ fastCircularPermutations }
\seealso{
	See \code{\link{}} and \code{\linkS4class{}} for reference.
	
	% Run \code{browseVignettes("filematrix")} for the list of vignettes.
}
