% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_cs.R
\name{interpolate_cs}
\alias{interpolate_cs}
\title{Interpolate between two hex colours}
\usage{
interpolate_cs(
  colour1 = "#ff7f23",
  colour2 = "#237f00",
  amt = 0.5,
  cs = "oklch"
)
}
\arguments{
\item{colour1}{colour to interpolate from}

\item{colour2}{colour to interpolate to}

\item{amt}{amount 0, 1 0 is all colour 1, 1 is all colour 2}

\item{cs}{colourspace to use}
}
\value{
hex colour
}
\description{
Interpolate between two hex colours
}
