% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_hue_lch.R
\name{shift_hue_cs}
\alias{shift_hue_cs}
\title{Shift hue in lch colourspace}
\usage{
shift_hue_cs(col = "#fdb975", shift = 30, cs = "oklch")
}
\arguments{
\item{col}{hex colour}

\item{shift}{amount to shift between 0 and 360}

\item{cs}{which colour space to use}
}
\value{
hex colour
}
\description{
Shift hue in lch colourspace
}
\examples{
shift_hue_cs()
}
