\name{shiftR-package}
\alias{shiftR-package}
\docType{package}
\title{
 Very Fast Circular Permutation Analysis on Binary Outcomes
}
\description{
 Very fast circular permutation analysis on binary outcomes
    using bitwise operations on binary representation of the data.
}
\details{
	\tabular{ll}{
		Package: \tab shiftR \cr
		Type: \tab Package \cr
		License: \tab LGPL-3 \cr 
		Depends: \tab methods \cr
	}
}
%\note{}
\author{
	Andrey A Shabalin \email{ashabalin@vcu.edu}
}
\keyword{ shiftR }
\seealso{
	See the main function \code{\link{circularPermutationAnalysis}} for more info.
	% Run \code{browseVignettes("filematrix")} for the list of vignettes.
}
