\name{shiftR-package}
\alias{shiftR-package}
\docType{package}
\title{
    Fast Enrichment Analysis via Circular Permutations of binary outcomes
}
\description{
    Fast enrichment analysis for locally correlated statistics
    via circular permutations.
    The analysis can be performed at multiple significance thresholds
    for both primary and auxiliary data sets with 
    with efficient correction for multiple testing.
}
\details{
    \tabular{ll}{
        Package: \tab shiftR \cr
        Type: \tab Package \cr
        License: \tab LGPL-3 \cr 
        Depends: \tab methods \cr
    }
}
\author{
	Andrey A Shabalin \email{ashabalin@vcu.edu}
}
\keyword{ shiftR }
\seealso{
	See the main function \code{\link{circularPermutationAnalysis}} for more info.
	% Run \code{browseVignettes("filematrix")} for the list of vignettes.
}
